refactor: 위시 수정 - layout으로 페이지 진입 권한 확인 분리
